+++ spring-ledger/src/main/java/com/noobcodezz/spring_ledger/models/LoanDto.java
@@ -3,6 +3,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 @Data
@@ -12,8 +13,8 @@ public class LoanDto {
     private String loanID;
     private String UserRefID;
     private String BankRefID;
-    private double principal;
-    private double roi;
+    private BigDecimal principal;
+    private BigDecimal roi;
     private int years;
     private Instant createdAt;
 }
